@@ -7,9 +7,10 @@ import android.content.Intent;
 import android.content.IntentFilter;
 import android.os.Bundle;
 import android.os.Parcelable;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.v4.content.LocalBroadcastManager;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.localbroadcastmanager.content.LocalBroadcastManager;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -29,7 +30,9 @@ public class PermissionApplyActivity extends Activity {
     private static final String KEY_PERMISSION_DENIED = "KEY_PERMISSION_DENIED";
     private static final String KEY_PERMISSION_DENIED_LIST = "KEY_PERMISSION_DENIED_LIST";
 
-    static void startAction(@NonNull Context context, @NonNull String[] permissions, @NonNull final PermissionManager.PermissionsListener listener) {
+    static void startAction(@NonNull Context context,
+                            @NonNull String[] permissions,
+                            @NonNull final PermissionManager.PermissionsListener listener) {
         Intent intent = new Intent(context, PermissionApplyActivity.class);
         intent.putExtra(PERMISSION_KEY, permissions);
         if (!(context instanceof Activity)) {
@@ -45,7 +48,8 @@ public class PermissionApplyActivity extends Activity {
                     if (intent.hasExtra(KEY_PERMISSION_GRANTED)) {
                         listener.onPermissionGranted();
                     } else if (intent.hasExtra(KEY_PERMISSION_DENIED)) {
-                        List<PermissionManager.NoPermission> noPermissionsList = intent.getParcelableArrayListExtra(KEY_PERMISSION_DENIED_LIST);
+                        List<PermissionManager.NoPermission> noPermissionsList =
+                                intent.getParcelableArrayListExtra(KEY_PERMISSION_DENIED_LIST);
                         listener.onPermissionDenied(noPermissionsList);
                     }
                 }
@@ -71,7 +75,8 @@ public class PermissionApplyActivity extends Activity {
                 super.onPermissionDenied(noPermissionsList);
                 Intent intent = new Intent(ACTION_PERMISSION);
                 intent.putExtra(KEY_PERMISSION_DENIED, true);
-                intent.putParcelableArrayListExtra(KEY_PERMISSION_DENIED_LIST, (ArrayList<? extends Parcelable>) noPermissionsList);
+                intent.putParcelableArrayListExtra(KEY_PERMISSION_DENIED_LIST,
+                        (ArrayList<? extends Parcelable>) noPermissionsList);
                 LocalBroadcastManager.getInstance(PermissionApplyActivity.this).sendBroadcast(intent);
                 finish();
             }
