@@ -3,13 +3,16 @@ package com.ysbing.ypermission.checker;
 import android.content.Context;
 import android.database.Cursor;
 import android.provider.ContactsContract;
-import android.support.annotation.NonNull;
+
+import androidx.annotation.NonNull;
 
 class ContactsReadTest {
 
     static boolean check(@NonNull Context context) {
         String[] projection = new String[]{ContactsContract.Data._ID, ContactsContract.Data.DATA1};
-        Cursor cursor = context.getContentResolver().query(ContactsContract.CommonDataKinds.Phone.CONTENT_URI, projection, null, null, null);
+        Cursor cursor = context.getContentResolver()
+                .query(ContactsContract.CommonDataKinds.Phone.CONTENT_URI, projection,
+                        null, null, null);
         if (cursor != null) {
             try {
                 PermissionTest.CursorTest.read(cursor);
