@@ -6,7 +6,8 @@ import android.hardware.Sensor;
 import android.hardware.SensorEvent;
 import android.hardware.SensorEventListener;
 import android.hardware.SensorManager;
-import android.support.annotation.NonNull;
+
+import androidx.annotation.NonNull;
 
 class SensorsTest {
 
